+++ src/custom-types/escape-path-with-spaces.ts
@@ -1,0 +1,1 @@
+declare module 'escape-path-with-spaces';