+++ src/renderer/components/PropertyList.tsx
@@ -54,7 +54,7 @@ export default class PropertyList extends React.Component<PropertyListProps> {
 
   public render() {
     return (
-      <div className="itemBar" ref={this.ref} style={{ width: 200, overflowY: 'scroll', paddingLeft: 8, borderLeft: '1px solid #999' }}>
+      <div className="itemBar" ref={this.ref} style={{ width: 225, overflowY: 'scroll', paddingLeft: 8, borderLeft: '1px solid #999' }}>
         {this.props.selected && this.makeControls(this.props.selected)}
       </div>
     );